LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/cssom-view/range-bounding-client-rect-with-display-contents-expected.txt

spacer before
HEIGHT: 30px
HEIGHT: 30px
spacer after

FAIL the space between elements using a range should be the same as using another method assert_equals: range.getBoundingClientRect().height expected 49 but got 60
